+++ homeassistant/components/zone/const.py
@@ -3,3 +3,5 @@
 CONF_PASSIVE = 'passive'
 DOMAIN = 'zone'
 HOME_ZONE = 'home'
+ATTR_PASSIVE = 'passive'
+ATTR_RADIUS = 'radius'
